Detail view from recommendation. Added 2 VCs for marthon list.
