+++ Movie Marathoner/Controllers/ViewControllers/WatchListViewController.swift
@@ -1,0 +1,47 @@
+//
+//  WatchListViewController.swift
+//  Movie Marathoner
+//
+//  Created by Andrew Saeyang on 2/2/22.
+//
+
+import UIKit
+
+class WatchListViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
+
+    // MARK: - Outlets
+    
+    @IBOutlet weak var tableView: UITableView!
+    
+    // MARK: - Properties
+    private let cellID = "watchListCell"
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        tableView.dataSource = self
+        tableView.delegate = self
+        
+        // Do any additional setup after loading the view.
+    }
+    
+    // MARK: - UITableViewSource
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 0
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        return UITableViewCell() // TODO: MAKE A CELL
+    }
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
